refactored main, did some other fixes too.
